# docs/arquitectura/contratos/openapi-incentivos.yaml
openapi: 3.0.3
info:
  title: DonaTrack - Incentivos Service API
  version: 1.0.0
  description: API REST para gamificación, misiones por Template Method, asignación de insignias y cómputo de ranking mensual de donantes.
servers:
  - url: http://localhost:8082
    description: Servidor local de desarrollo
paths:
...
  /api/incentivos/donantes/{donanteId}:
    post:
      summary: Registrar un nuevo donante en el sistema de incentivos
      operationId: registrarDonanteIncentivos
      parameters:
        - name: donanteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrarDonanteRequest'
      responses:
        '201':
          description: Donante registrado exitosamente en incentivos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonanteRegistradoDTO'
        '400':
          description: Datos inválidos o discrepancia de ID
    delete:
      summary: Dar de baja donante en incentivos
      operationId: darDeBajaDonanteIncentivos
      parameters:
        - name: donanteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Donante dado de baja exitosamente
        '404':
          description: Donante no encontrado
    patch:
      summary: Modificar atributos de donante en incentivos
      operationId: modificarDonanteIncentivos
      parameters:
        - name: donanteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModificarDonanteRequest'
      responses:
        '200':
          description: Donante modificado exitosamente
        '400':
          description: Payload inválido
        '404':
          description: Donante no encontrado
    get:
      summary: Consultar perfil consolidado del donante en incentivos
      operationId: obtenerDonantePerfil
      parameters:
        - name: donanteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Perfil consolidado del donante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DonantePerfilDTO'
        '404':
          description: Donante no encontrado
  /api/incentivos/donantes/{donanteId}/ascensos:
    get:
      summary: Consultar historial de ascensos de categoría del donante
      operationId: obtenerAscensosDonante
      parameters:
        - name: donanteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Historial de ascensos de categoría
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CambioCategoriaDTO'
        '404':
          description: Donante no encontrado
...
components:
  schemas:
...
    RegistrarDonanteRequest:
      type: object
      required:
        - idDonante
        - idPersona
        - nombre
      properties:
        idDonante:
          type: string
          format: uuid
        idPersona:
          type: string
          format: uuid
        nombre:
          type: string
          minLength: 2
          maxLength: 100
    DonanteRegistradoDTO:
      type: object
      properties:
        donanteId:
          type: string
          format: uuid
        categoria:
          type: string
    ModificarDonanteRequest:
      type: object
      required:
        - nombre
      properties:
        nombre:
          type: string
          minLength: 2
          maxLength: 100
...
    DonantePerfilDTO:
      type: object
      properties:
        donanteId:
          type: string
          format: uuid
        idPersona:
          type: string
          format: uuid
        nombre:
          type: string
        categoria:
          type: string
          enum: [COLABORADOR, SOSTENEDOR, TRANSFORMADOR]
        fechaRegistro:
          type: string
          format: date
        misionesCompletadas:
          type: integer
        insigniasGanadas:
          type: integer
    CambioCategoriaDTO:
      type: object
      properties:
        categoriaAnterior:
          type: string
          enum: [COLABORADOR, SOSTENEDOR, TRANSFORMADOR]
        categoriaNueva:
          type: string
          enum: [COLABORADOR, SOSTENEDOR, TRANSFORMADOR]
        fecha:
          type: string
          format: date